Rename memory store module
