+++ p2panda-store/src/memory.rs
